Renamed java 'custom' test to 'customSequence'.
