+++ java/test/src/main/java/test/Ice/customSequence/TestI.java
@@ -1,14 +1,14 @@
 // Copyright (c) ZeroC, Inc.
 
-package test.Ice.custom;
+package test.Ice.customSequence;
 
 import com.zeroc.Ice.Communicator;
 import com.zeroc.Ice.Current;
 
-import test.Ice.custom.Test.A;
-import test.Ice.custom.Test.E;
-import test.Ice.custom.Test.S;
-import test.Ice.custom.Test.TestIntf;
+import test.Ice.customSequence.Test.A;
+import test.Ice.customSequence.Test.E;
+import test.Ice.customSequence.Test.S;
+import test.Ice.customSequence.Test.TestIntf;
 
 import java.nio.ByteBuffer;
 import java.nio.DoubleBuffer;
